Module error_stack::future

Extension for convenient usage of Reports returned by Future s.

Extends Future with the same methods as ResultExt but calls the methods on polling.

Structs

FutureWithAttachment

Adaptor returned by FutureExt::attach.

FutureWithContext

Adaptor returned by FutureExt::change_context.

FutureWithLazyAttachment

Adaptor returned by FutureExt::attach_lazy.

FutureWithLazyContext

Adaptor returned by FutureExt::change_context_lazy.

FutureWithLazyPrintableAttachment

Adaptor returned by FutureExt::attach_printable_lazy.

FutureWithPrintableAttachment

Adaptor returned by FutureExt::attach_printable.

Traits

FutureExt

Extension trait for Future to provide contextual information on Reports.