error_stack

Struct Frame

pub struct Frame { /* private fields */ }

A single context or attachment inside of a Report.

Frames are organized as a singly linked list, which can be iterated by calling Report::frames(). The head contains the current context or attachment, and the tail contains the root context created by Report::new(). The next Frame can be accessed by requesting it by calling Report::request_ref().

Implementations

impl Frame

pub const fn sources(&self) -> &[Self]

Returns a shared reference to the source of this Frame.

This corresponds to the Frame below this one in a Report.

pub fn sources_mut(&mut self) -> &mut [Self]

Returns a mutable reference to the sources of this Frame.

This corresponds to the Frame below this one in a Report.

pub fn kind(&self) -> FrameKind<'_>

Returns how the Frame was created.

pub fn request_ref<T>(&self) -> Option<&T>
where T: ?Sized + 'static,

Available on nightly only.

Requests the reference to T from the Frame if provided.

pub fn request_value<T>(&self) -> Option<T>
where T: 'static,

Available on nightly only.

Requests the value of T from the Frame if provided.

pub fn is<T: Send + Sync + 'static>(&self) -> bool

Returns if T is the held context or attachment by this frame.

pub fn downcast_ref<T: Send + Sync + 'static>(&self) -> Option<&T>

Downcasts this frame if the held context or attachment is the same as T.

pub fn downcast_mut<T: Send + Sync + 'static>(&mut self) -> Option<&mut T>

Downcasts this frame if the held context or attachment is the same as T.

pub fn type_id(&self) -> TypeId

Returns the TypeId of the held context or attachment by this frame.