Trait error_stack::IntoReportCompat

pub trait IntoReportCompat: Sized {
    type Ok;
    type Err;

    // Required method
    fn into_report(self) -> Result<Self::Ok, Report<Self::Err>>;
}

Compatibility trait to convert from external libraries to Report.

Note: It’s not possible to implement Context on other error libraries’ types from within error-stack as the trait has a blanket implementation relying on Error. Thus, implementing the trait would violate the orphan rule; the upstream crate could implement Error and this would imply an implementation for Context. This also implies, that it’s not possible to implement ResultExt from within error-stack.

Required Associated Types

type Ok

Type of the Ok value in the Result

type Err

Type of the resulting Err variant wrapped inside a Report<E>.

Required Methods

fn into_report(self) -> Result<Self::Ok, Report<Self::Err>>

Converts the Err variant of the Result to a Report

Object Safety

This trait is not object safe.